% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craio.R
\name{craio}
\alias{craio}
\title{craio}
\usage{
craio(l)
}
\arguments{
\item{l}{numeric. The parameter to generate a random index}
}
\value{
The generated full name.
}
\description{
A Caio's random last name generator.
}
\examples{
{
\donttest{
  craio(5)
}
}
}
